+++ backend/app/data_products/service.py
@@ -1,5 +1,5 @@
 import copy
-from typing import Optional, Sequence
+from typing import Sequence, assert_never
 from uuid import UUID
 from warnings import deprecated
 
@@ -14,7 +14,7 @@ from app.abstract_data_product.input_ports.model import (
     InputPort as InputPortModel,
 )
 from app.abstract_data_product.service import AbstractDataProductService
-from app.authorization.role_assignments.enums import DecisionStatus
+from app.authorization.role_assignments.enums import AssignmentFilter, DecisionStatus
 from app.authorization.roles.schema import Prototype
 from app.authorization.service import DATA_PRODUCT_READER_ROLE
 from app.configuration.data_product_lifecycles.model import (
@@ -135,7 +135,9 @@ class DataProductService(AbstractDataProductService):
 
     def get_data_products(
         self,
-        filter_to_user_with_assigment: Optional[UUID] = None,
+        *,
+        current_user: User,
+        assignment_filter: AssignmentFilter,
     ) -> Sequence[DataProductModel]:
         default_lifecycle = self.db.scalar(
             select(DataProductLifeCycleModel).filter(
@@ -146,22 +148,31 @@ class DataProductService(AbstractDataProductService):
             selectinload(DataProductModel.tags).raiseload("*"),
             undefer(DataProductModel.input_port_count),
         )
-        if filter_to_user_with_assigment:
-            query = query.filter(
-                DataProductModel.assignments.any(
-                    user_id=filter_to_user_with_assigment,
-                    decision=DecisionStatus.APPROVED,
+        match assignment_filter:
+            case AssignmentFilter.ALL:
+                pass
+            case AssignmentFilter.ONLY_ASSIGNED:
+                query = query.filter(
+                    DataProductModel.assignments.any(
+                        user_id=current_user.id,
+                        decision=DecisionStatus.APPROVED,
+                    )
                 )
-            )
+            case _:
+                assert_never(assignment_filter)
         query = query.order_by(asc(DataProductModel.name))
 
         dps = self.db.scalars(query).unique().all()
 
+        auth = Authorization()
+        filtered = []
         for dp in dps:
-            if not dp.lifecycle:
-                dp.lifecycle = default_lifecycle
+            if auth.has_read_access_to_data_product(current_user, dp):
+                if not dp.lifecycle:
+                    dp.lifecycle = default_lifecycle
+                filtered.append(dp)
 
-        return dps
+        return filtered
 
     def get_owners(self, id: UUID) -> Sequence[User]:
         data_product = ensure_data_product_exists(